Rename the message-notification cron to messagenotifications
