+++ backend/service/cron/__init__.py
@@ -6,7 +6,7 @@ from service.cron.clubseo import (
     refresh_club_overlap_forever,
 )
 from service.cron.garbagerecords import delete_garbage_records_forever
-from service.cron.notifications import send_notifications_forever
+from service.cron.messagenotifications import send_notifications_forever
 from service.cron.nsfwphotorunner import predict_nsfw_photos_forever
 from service.cron.visitornotifications import (
     send_visitor_notifications_forever,
